refactor: extract config/shader management into minarch_config.c
Move Config_init/quit/load/free/readOptions/readControls/write/restore,
Config_syncFrontend/syncShaders, applyShaderSettings, initShaders,
list_files_in_folder, setOverclock, updateCPUMonitor, and
getScreenScalingDesc/Count out of minarch.c into dedicated module.

Rename static screenx/screeny to cfg_screenx/cfg_screeny to avoid
multiple-definition conflict with generic_video.c in common/.
